moved missing items calculation to missing items component
